From: Andreas Schulz <ats@g386bsd.first.gmd.de>
I have a question to the current default of the gethostnamadr.c for
name-resolution. The new behaviour is different from the old behaviour.
The old behaviour was to ask the bind and fallback to hosts if no bind
is running, the new behaviour is to only try bind(if you don't have
an /etc/host.conf file). This default makes it difficult to boot up
a networked machine, i have used the default /etc/netstart file that
calls "ifconfig" with a hostname, but that fails now.
Wouldn't it be better, to make the default SERVICE_BIND, SERVICE_HOST,
SERVICE_NONE or SERVICE_HOST, SERVICE_NONE instead of the current
SERVICE_BIND, SERVICE_NONE, if no /etc/host.conf file exists ?