From: bde@kralizec.zeta.org.au (Bruce Evans) Subject: Re:  slattach patches
if_sl.c was not calling the output process when it thinks the output
queue is empty.  But the output process sometimes need to clean up
internal state, at least for sio.  I was having trouble killing
slattach because it got stuck in close waiting for output to drain when
CRTSCTS was on but MSR_CTS was off.  stty -crtscts didn't work although
output was drained because comstart() wasn't called so the process
remained stuck in ttywait().

if_sl.c wasn't ignoring carrier when CLOCAL is set (bug reported ages
ago for BSDI).