>From bde@kralizec.zeta.org.au Tue Jul 13 08:41:26 1993 >Date: Tue, 13 Jul 93 05:07:45 EST
/bin/sh has incomplete and broken exec permission checking.  It did not
allow root to exec /sbin/disklabel (perms r-xr-s---, owner bin, group
operator) although the exec would have would if /bin/sh tried it.  This
patch fixes that.  Group access is still not supported.  I would prefer
to attempt the exec the first time instead of using stat().

>From bde@kralizec.zeta.org.au Sun Jul 18 10:31:04 1993
>Date: Sun, 18 Jul 93 04:10:26 EST

Here is a better version.  It also contains my previous fix which was for
another bug (this bug is masked by the getgroups fix because root is
usually in most groups).  I don't like all the ifdefs for POSIX support
but they seem to be necessary for portable code.