From: Gary Shaffer <grs@orcas.cbits.SAIC.Com> Subject: PING patch for FreeBSD-1.0
Here is the fix for ping so it returns 1 if no packets are returned.  I have
includes the complete "finish" function found in /usr/src/sbin/ping/ping.c
for your review.  The rational behind this fix is includes as a comment in
the code. changes start with the line " * 13NOV93 GRS ". Feel free do to
what you want with this fix..