From Rodney Volz:
In FreeBSD/ufs, files are always created with their group-id
set to the group-id of their parent-directory. This behaviour
is neither POSIX nor BSD-conforming.

Posix says, that the group-id of newly created files should be set to
the group-id of the creator. An optional Posix feature is, that
if the parent directory has a group S-bit set, the group-id
of the new file should be set to the group-id of its parent
directory.