Beginning of change set for making more friendly laptop configurations. Changes _only_ take effect if `options LAPTOP' is set.
This disables the update process from running if you're on a laptop, where you
do NOT want the bloody update to come along every few minutes and kick your
sleeping hard drive back to power-sucking life again.  Ideally, this should be
done with some sort of runtime configuration tool that tells update how often
to sync (with 0 == don't bother), but that's a lot more work than this was! :)