Add more segment registers to list of registers (fake for %fs and %gs). This might be useful for debugging applications that use a special LDT. However, printing of all segment registers is currently broken.
Don't print "last exception: " before the FPU opcode and pc, etc.  The
opcode and pc, etc., are for the last FPU _instruction_.

Pass the saved exception status word to print_387_status() so that the
exception(s) that caused or will cause a trap can be seen.  The kernel
has supported this since 1.1 or before.  The kernel still clobbers the
tag word if a trap occurs.

Remove unused null function clear_regs().


git-svn-id: svn+ssh://svn.freebsd.org/base/head@5319 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f