From "Philippe Charnier" <charnier@lirmm.fr>:
There is a bug in sh: the built in command "fc -l" generates
a core dump (*NULL in not_fcnumber).

According to the sh manual page (fc  -l [-nr] [first [last]]), fc -l
is a correct sequence (in that case, values are defaulted to -16 and -1)
but fails when first is not given.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@8289 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f