Updated comments.  The comments about the unused addresses get broken almost every time someone uses an address.  This file is probably not the right place to keep track of the unused addresses (or used addresses :->).
Fixed comments on #endif's to match code.

Added defines for ASC and GSC sizes.  This file is not the right place
to keep track of scanner addresses, but while there here and we
pretend to keep track of unused addresses, the sizes need to be here
too.

Sorted IO_*SIZE defines.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@12366 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f