Update bcopyb & bcopy to reflect changes I made in the libc version of bcopy:
Be smarter about handling overlapped copies and only go backwards if it
is really necessary. Going backwards on a P6 is much slower than forwards
and it's a little slower on a P5. Also moved the count mask and 'std'
down a few lines - it's a couple percent faster this way on a P5.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@13065 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f