Change the messages slightly when there is no "mount_type" executable found when the user specifies "mount -t type".  Instead of printing out one message for each path element (/sbin, /usr/sbin), it prints out:
mount: exec mount_type not found in /sbin, /usr/sbin: No such file or directory

The code is quite long for such a stupid little piece of aesthesism
but it is very straghtforward so I guess it's ok.  Besides, I don't
want to do a "char foo[100];" and have malloc break down when someone
decides to add a few more paths to a variable that's far apart from
this code. :)

By the way, there is no malloc() off-by-one error for the '\0' at the
end of the string although I don't explicitly add 1 to the length.
The code allocates strlen(path element)+2 bytes for each path element,
and doesn't use the last two bytes (for the delimiting ", ").

Reviewed by:	the list (I hope)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@14626 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f