Create a new flag, "nochange", which will specify that a directory is to be created if it's missing, otherwise completely ignore it's modes and owners.  Primary intended targets:  /usr/src and /usr/obj.
Adjust the 'not created: File exists' message to mention that it's a
directory that's the problem, otherwise it doesn't make sense.

I had created chown-style  -L and -P flag to control logical/physical mode
(ie: whether symlinks were followed), but the nochange flag is enough to
get the blasted thing out of my hair so I took them back out.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@36670 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f