The first stage of console driver reorganization: activate new keyboard and video card drivers.
Because of the changes, you are required to update your kernel
configuration file now!

The files in sys/dev/syscons are still i386-specific (but less so than
before), and won't compile for alpha and PC98 yet.

syscons still directly accesses the video card registers here and
there; this will be rectified in the later stages.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@42504 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f