Initialize the si_bsize fields of the new dev structure.  This fixes some swapper problems analogous to those experienced with ccd.
This fix is a kludge: since we currently don't track the "sector size"
in a volume label, we guess a worst case (4 kB, as used by vnode
devices).  If the concept of sector size is here to stay, I'll make
some changes to track the "sector size" of a volume.  This will
probably be the maximum of the sector sizes of all component drives,
but things could get ugly if we start allowing non-standard sector
sizes such as 524 bytes.

Unkludged-version-submitted-by:	phk


git-svn-id: svn+ssh://svn.freebsd.org/base/head@50537 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f