Add $FreeBSD$
Correct the date on which DST begins in Victoria and New South Wales
in the year 2000: as the result of some local function in Sydney, DST
will begin on Sunday, 27 August instead of Sunday, 29 Cotober.

This change had already been made, but the date was incorrectly
specified as (Saturday) 26 August.

Submitted by:	Howard Lowndes <lannet@lannet.com.au>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@52770 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f