o  Add games/caesar to the list of bootstrap-tools so that a    buildworld doesn't break because the host doesn't have any    games installed, o  Add a new build stage: TMAKE. TMAKE builds all the build-tools    targets in the respective makefiles. Note that these targets    don't use the bootstrap tools, o  Add elf2exe to the bootstrap-tools when cross-building Alpha on    other platforms, o  Add ${WORLDTMP}/usr/games to TMPPATH, o  Remove ${WORLDTMP}/usr/bin even when NOCLEAN is defined. This    prevents using any bootstrap-tools previously installed. Most    importantly, it prevents using the cross-compiler when we still    need the native compiler.
The current stages are BMAKE, TMAKE, XMAKE and IMAKE in that order.
BMAKE builds bootstrap-tools that either solve compatibility problems
      or are needed as cross-tools,
TMAKE builds the support tools necessary by some parts in the source
      tree and also performs the cleandir and par-obj targets,
XMAKE builds the includes, libraries and everything (resp.), and
IMAKE installs the world. This stage needs further work if it's to be
      used to install -current over -stable for example.

This is the last major update towards cross-building.