New function checkupdates: check whether configuration updates are enabled and print a warning message in varous places if they are not.
Fix typos in comments.

Change some indents to approach style(9).

initvol: If init fails, place the subdisks in a "down" state.

vinum_start(): If we're starting a plex, try doing it directly.  It's
possible that the state of the plex differs from that of the subdisks,
for example if somebody has used 'setupstate'.

New functions vinum_raid4 and vinum_raid5 to create RAID-4 and RAID-5
volumes, analagous to vinum_mirror and vinum_stripe.

vinum_checkparity: Don't try to check the parity of a non-parity plex,
print a rude remark instead.

Approved-by:  jkh


git-svn-id: svn+ssh://svn.freebsd.org/base/head@57604 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f